requirements.txt
crawl4ai[all]==0.4.247
fastapi>=0.104.1
uvicorn>=0.24.0
playwright>=1.41.0
psutil>=5.9.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.5.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
markdown>=3.5.0 